Update filtering for sysadmins and correct docs
README.md contains a few references to the default value of the
node['nagios']['users_databag_group'] setting as 'sysadmins' when
the code actually references the group as 'sysadmin'. Corrected
the docs to match the code to alleviate any confusion.

In server.rb, we do a search in the users databag to find user
items that match that group. However, if people are also using the
user cookbook to manage users, they may have some user items with
an action of ":remove". These users should be excluded from this
search so they will not get added to the htpasswd file, OpenID
authorization, etc., which I have done via adding the following
to the query:

  NOT action:remove

If the action value is not specified in the item, then it does not
fail, so everything is fine-- this only excludes users who should
be removed.